- fix version dependencies - fixing imports - upgrade target and compile sdk versions of android to 34 - rename files - now use underscores - small fixes and improvements
