+++ .gitignore
@@ -44,6 +44,3 @@ app.*.map.json
 /android/app/release
 
 /lib/auth_token.dart
-
-#Custom
-pubspec.lock
