+++ lib/features/appConfiguration/entity/build_configuration.dart
