+++ lib/main.dart
@@ -2,9 +2,9 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:intl/date_symbol_data_local.dart';
-import 'package:zachranobed/features/appConfiguration/AppConfiguration.dart';
-import 'package:zachranobed/features/appConfiguration/mapper/AppConfigurationMapper.dart';
-import 'package:zachranobed/features/login/di/LoginDependencyContainer.dart';
+import 'package:zachranobed/features/appConfiguration/app_configuration.dart';
+import 'package:zachranobed/features/appConfiguration/mapper/app_configuration_mapper.dart';
+import 'package:zachranobed/features/login/di/login_dependency_container.dart';
 import 'package:zachranobed/firebase/firebase_options.dart';
 import 'package:zachranobed/firebase/notifications.dart';
 import 'package:zachranobed/services/ioc_container.dart';
