+++ lib/routes/app_router.gr.dart
@@ -11,7 +11,7 @@
 import 'package:auto_route/auto_route.dart' as _i12;
 import 'package:cloud_firestore/cloud_firestore.dart' as _i17;
 import 'package:flutter/material.dart' as _i13;
-import 'package:zachranobed/features/debug/DebugScreen.dart' as _i3;
+import 'package:zachranobed/features/debug/debug_screen.dart' as _i3;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
     as _i7;
 import 'package:zachranobed/models/box_movement.dart' as _i14;
